Move `ipc_service` to `firezone-gui-client`
