+++ rust/gui-client/src-tauri/src/clear_logs.rs
