+++ rust/gui-client/src-tauri/src/ipc/linux.rs
@@ -87,6 +87,7 @@ impl Server {
 fn ipc_path(id: ServiceId) -> PathBuf {
     match id {
         ServiceId::Prod => PathBuf::from("/run").join(BUNDLE_ID).join("ipc.sock"),
+        #[cfg(test)]
         ServiceId::Test(id) => firezone_bin_shared::known_dirs::runtime()
             .expect("`known_dirs::runtime()` should always work")
             .join(format!("ipc_test_{id}.sock")),
