+++ rust/gui-client/src-tauri/src/ipc/macos.rs
