+++ rust/gui-client/src-tauri/src/service/linux.rs
@@ -1,14 +1,15 @@
-use super::Cli;
+use std::path::PathBuf;
+
 use anyhow::{Result, bail};
-use firezone_bin_shared::signals;
+use firezone_bin_shared::{DnsControlMethod, signals};
 
 use firezone_telemetry::Telemetry;
 
 /// Cross-platform entry point for systemd / Windows services
 ///
 /// Linux uses the CLI args from here, Windows does not
-pub(crate) fn run_ipc_service(cli: Cli) -> Result<()> {
-    let (_handle, log_filter_reloader) = super::setup_logging(cli.log_dir)?;
+pub fn run_ipc_service(log_dir: Option<PathBuf>, dns_control: DnsControlMethod) -> Result<()> {
+    let (_handle, log_filter_reloader) = crate::logging::setup_ipc(log_dir)?;
     if !elevation_check()? {
         bail!("IPC service failed its elevation check, try running as admin / root");
     }
@@ -20,7 +21,7 @@ pub(crate) fn run_ipc_service(cli: Cli) -> Result<()> {
     let mut telemetry = Telemetry::default();
 
     rt.block_on(super::ipc_listen(
-        cli.dns_control,
+        dns_control,
         &log_filter_reloader,
         &mut signals,
         &mut telemetry,
@@ -36,10 +37,10 @@ pub(crate) fn run_ipc_service(cli: Cli) -> Result<()> {
 /// Returns true if the IPC service can run properly
 // Fallible on Windows
 #[expect(clippy::unnecessary_wraps)]
-pub(crate) fn elevation_check() -> Result<bool> {
+pub fn elevation_check() -> Result<bool> {
     Ok(nix::unistd::getuid().is_root())
 }
 
-pub(crate) fn install_ipc_service() -> Result<()> {
+pub fn install_ipc_service() -> Result<()> {
     bail!("`install_ipc_service` not implemented and not needed on Linux")
 }
