+++ rust/gui-client/src-tauri/src/service/macos.rs
@@ -1,0 +1,18 @@
+use anyhow::{Result, bail};
+use firezone_bin_shared::DnsControlMethod;
+use std::path::PathBuf;
+
+pub fn run_ipc_service(log_dir: Option<PathBuf>, _dns_control: DnsControlMethod) -> Result<()> {
+    // We call this here to avoid a dead-code warning.
+    let (_handle, _log_filter_reloader) = crate::logging::setup_ipc(log_dir)?;
+
+    bail!("not implemented")
+}
+
+pub fn elevation_check() -> Result<bool> {
+    bail!("not implemented")
+}
+
+pub fn install_ipc_service() -> Result<()> {
+    bail!("not implemented")
+}
