+++ rust/headless-client/Cargo.toml
@@ -51,13 +51,12 @@ libc = { workspace = true }
 nix = { workspace = true, features = ["fs", "user", "socket"] }
 resolv-conf = { workspace = true }
 rtnetlink = { workspace = true }
-sd-notify = "0.4.5" # This is a pure Rust re-implementation, so it isn't vulnerable to CVE-2024-3094
+sd-notify = { workspace = true }
 
 [target.'cfg(target_os = "windows")'.dependencies]
 ipconfig = "0.3.2"
 itertools = { workspace = true }
 known-folders = { workspace = true }
-windows-service = "0.8.0"
 winreg = { workspace = true }
 
 [target.'cfg(windows)'.dependencies.windows]
