--- rust/headless-client/src/ipc_service/macos.rs
@@ -1,17 +1,0 @@
-use super::Cli;
-use anyhow::{Result, bail};
-
-pub(crate) fn run_ipc_service(cli: Cli) -> Result<()> {
-    // We call this here to avoid a dead-code warning.
-    let (_handle, _log_filter_reloader) = super::setup_logging(cli.log_dir)?;
-
-    bail!("not implemented")
-}
-
-pub(crate) fn elevation_check() -> Result<bool> {
-    bail!("not implemented")
-}
-
-pub(crate) fn install_ipc_service() -> Result<()> {
-    bail!("not implemented")
-}
